fix(checker): preserve canonical assignment flow
